Add global memory allocator for C++
Note that this tap.h is Firebirds slightly modified version of the stock
Topfield tap.h. You should replace the stock tap.h file with this one.

This tiny commit adds global new() & delete() so that C++ taps use the
tap memory allocators. This commit really should be also be added to
Firebirds version, and ultimately to the stock tap.h.